Move rslist package to ui.posts_rs
