+++ WordPress/src/main/AndroidManifest.xml
@@ -960,7 +960,7 @@
             android:theme="@style/WordPress.NoActionBar"/>
 
         <activity
-            android:name=".ui.posts.rslist.PostRsListActivity"
+            android:name=".ui.posts_rs.PostRsListActivity"
             android:label="@string/my_site_btn_blog_posts"
             android:launchMode="singleTop"
             android:theme="@style/WordPress.NoActionBar"/>
