+++ WordPress/src/main/java/org/wordpress/android/ui/posts_rs/PostRsListUiState.kt
@@ -1,6 +1,6 @@
-package org.wordpress.android.ui.posts.rslist
+package org.wordpress.android.ui.posts_rs
 
-import org.wordpress.android.ui.posts.rslist.models.PostRsModel
+import org.wordpress.android.ui.posts_rs.models.PostRsModel
 import org.wordpress.android.util.DateTimeUtils
 import org.wordpress.android.util.DateTimeUtilsWrapper
 import java.text.DateFormat
